Update checksums with transferred files only in non dry-run mode

+++ sumo-remote.sh
@@ -12,6 +12,9 @@ del REMOTE & Remove a remote REMOTE
 list REMOTE & List avilable REMOTES
 push REMOTE & Push files which are available locally but missing on REMOTE to REMOTE
 pull REMOTE & Pull files which are available on REMOTE but missing locally from REMOTE
+common REMOTE & List files which are present locally and in REMOTE
+local-only REMOTE & List files which are present locally but not in REMOTE
+remote-only REMOTE & List files which are present in REMOTE but not locally
 &
 OPTION:
 -r | --repository REPOSITORY & Use REPOSITORY for storing checksums
@@ -57,8 +60,10 @@ function rsync_transfer() {
     comm -2 -3 "${ROOT}/remotes/${src_remote}/checksums" "${ROOT}/remotes/${dst_remote}/checksums" > "$TEST_TEMP_DIR/transferred"
     rsync -Prhv $DRY_RUN --files-from <(select_filenames < "$TEST_TEMP_DIR/transferred") "$src_path" "$dst_path"
 
-    cat "${ROOT}/remotes/${dst_remote}/checksums" "$TEST_TEMP_DIR/transferred" > "$TEST_TEMP_DIR/wip"
-    commit "${ROOT}/remotes/${dst_remote}/checksums"
+    if [ -n "$DRY_RUN" ]; then
+        cat "${ROOT}/remotes/${dst_remote}/checksums" "$TEST_TEMP_DIR/transferred" > "$TEST_TEMP_DIR/wip"
+        commit "${ROOT}/remotes/${dst_remote}/checksums"
+    fi
 }
 
 case "$1" in
